requirements.txt
azure-functions==1.23.0
azure-identity==1.19.0
requests==2.32.3
pyodbc==5.2.0
